Rename client.py to bitcoin_price_client.py and align test filename
